DatAgora demo finalised
- Added own image repo in demo folder
- Removed DatAgora BaseDemo
- demo files modified to comply with new specifications
