+++ UD-Viz-Core/examples/Demo_DatAgora_PartDieu/Demo.css
@@ -1,221 +1,4 @@
-a {
-    color : lightgreen;
-}
-
-* {
-    z-index: inherit;
-}
-
-html {
-    font-family: sans-serif;
-    font-size: 14px;
-    height: 100%;
-    width: 100%;
-    overflow: hidden;
-}
-
-body {
-    margin: 0;
-    background-color: black;
-    height: 100%;
-}
-
-#viewerDiv {
-    height: 100%;
-    position: relative;
-}
-
-canvas {
-    position: absolute;
-}
-
-#_base_demo {
-    display: grid;
-    height: 100%;
-    grid-template-rows: 50px auto;
-}
-
-header {
-    background-color: #333333;
-    display: grid;
-    grid-template-columns: auto auto;
-}
-
-header * {
-    vertical-align: top;
-}
-
-header h1 {
-    display: inline-block;
-    color: white;
-    margin: 5px;
-    font-size: 28px;
-    height: 100%;
-}
-
-#_base_demo_struct_header_panel {
-    text-align: right;
-    padding-right: 10px;
-}
-
-#_base_demo_stuct_main_panel {
-    display: grid;
-    grid-template-columns: 60px auto;
-}
-
-nav {
-    background-color: #222222;
-    display: inline-block;
-    height: 100%;
-    z-index: 50001;
-}
-
-section {
-    position: relative;
-    height: 100%;
-}
-
-#openSidebar {
-    display: none;
-}
-
-#logoIMU, #logoLIRIS,#logoUnivLyon,#logoGrandLyon {
+#logoUnivLyon,#logoGrandLyon {
     height: 40px;
     margin: 5px 10px 5px 5px;
-}
-
-.choiceMenu:hover, #openSidebar:hover, #openHamburger:hover, #closeHamburger:hover {
-    cursor: pointer;
-}
-
-#_base_demo_auth_frame_location {
-    display: inline-block;
-}
-
-#_base_demo_profile {
-    
-}
-
-#_base_demo_profile_name {
-    display: inline-block;
-    color: white;
-    font-size: 20px;
-    padding: 10px;
-}
-
-#_base_demo_profile_name em {
-    font-style: normal;
-    font-weight: bold;
-    font-variant: small-caps;
-}
-
-.logInOut {
-    display: inline-block;
-    border: 0;
-    background: 0;
-    color : lightgreen;
-    padding: 11px;
-    font-size: 18px;
-    text-decoration: underline;
-    font-style: italic;
-}
-
-#loginRegistration {
-    left: 23%;
-}
-
-.logInOut:hover {
-    cursor: pointer;
-}
-
-#_base_demo_menu {
-    list-style: none;
-    padding: 0;
-}
-
-.choiceMenu {
-    position: relative;
-    height: 60px;
-}
-
-._base_demo_menu_hint {
-    position: absolute;
-    background-color: #666666;
-    z-index: 50000;
-    height: 100%;
-    padding: 20px;
-    padding-top: 12px;
-    font-size: 24px;
-    display: inline-block;
-    margin: 0;
-    left: 100%;
-    top: 0;
-    white-space: nowrap;
-    box-sizing: border-box;
-    color: white;
-    visibility: hidden;
-    opacity: 0;
-    text-shadow: 1px 1px 1px #0009;
-    box-shadow: 1px 0px 1px -1px #0009, 0px 1px 1px -1px #0009, 0px -1px 1px -1px #0009;
-}
-
-.choiceMenu:hover > ._base_demo_menu_hint {
-    visibility: visible;
-    opacity: 1;
-}
-
-.choiceMenuSelected > ._base_demo_menu_hint {
-    background-color: #f09f45;
-}
-
-#activateTemporal:checked ~ header #temporalMenu {
-    background-color: rgba(255, 255, 255, 0.1);
-}
-
-.choiceMenuSelected::after {
-    position: absolute;
-    height: 100%;
-    right: 0;
-    top: 0;
-    width: 5px;
-    background-color: #f09f45;
-    content: "";
-    z-index: inherit;
-}
-
-.nonVisible{
-    display: none;
-}
-
-#_base_demo_menu .choiceMenu:hover {
-    background-color: #666666;
-}
-
-.menuIcon {
-    display: inline-block;
-    width: 36px;
-    margin: 12px 12px 12px 12px;
-    flex-shrink: 0;
-    opacity: 0.9;
-}
-
-#openSidebar:checked ~ #_base_demo_menu {
-    display: block;
-    transform: translateX(0);
-}
-
-#temporal {
-    display: none;
-}
-
-#activateTemporal:checked ~ #temporal {
-    display: block;
-}
-
-#researchContainer {
-    top: 50px;
-}
-
-#docBrowserWindow {
-    top: 50px;
 }
