+++ UD-Viz-Core/examples/Demo_DatAgora_PartDieu/DemoConfigData.json
@@ -23,6 +23,30 @@
       "max_y": "5175850"
     }
   },
+    
+  "logo_style" : {
+    "height" : "40px",
+    "margin" : "5px 10px 5px 5px"
+  },
+
+  "projection":"EPSG:3946",
+  "background_image_layer":
+  {
+    "url":"https://download.data.grandlyon.com/wms/grandlyon",
+    "name":"Ortho2018_Dalle_unique_8cm_CC46",
+    "version":"1.3.0",
+    "format":"image/jpeg",
+    "layer_name":"Base_Map",
+    "transparent":true
+  },
+  "elevation_layer":
+  {
+    "url":"https://download.data.grandlyon.com/wms/grandlyon",
+    "name":"MNT2018_Altitude_2m",
+    "format":"image/jpeg",
+    "layer_name":"wms_elevation_test"
+  },
+
   "camera": {
     "lyon_part_dieu":{
       "position": {
